+++ engine/src/flutter/shell/testing/tester_main.cc
@@ -30,86 +30,14 @@
 #include "third_party/dart/runtime/include/dart_api.h"
 #include "third_party/skia/include/core/SkSurface.h"
 
-// Impeller should only be enabled if the Vulkan backend is enabled.
-#define ALLOW_IMPELLER (IMPELLER_SUPPORTS_RENDERING && IMPELLER_ENABLE_VULKAN)
+#include "flutter/shell/testing/tester_vk.h"
 
-#if ALLOW_IMPELLER
-#include <vulkan/vulkan.h>                                        // nogncheck
-#include "impeller/display_list/aiks_context.h"                   // nogncheck
-#include "impeller/entity/vk/entity_shaders_vk.h"                 // nogncheck
-#include "impeller/entity/vk/framebuffer_blend_shaders_vk.h"      // nogncheck
-#include "impeller/entity/vk/modern_shaders_vk.h"                 // nogncheck
+#if TESTER_ENABLE_VULKAN
 #include "impeller/renderer/backend/vulkan/context_vk.h"          // nogncheck
 #include "impeller/renderer/backend/vulkan/surface_context_vk.h"  // nogncheck
 #include "impeller/renderer/context.h"                            // nogncheck
-#include "impeller/renderer/vk/compute_shaders_vk.h"              // nogncheck
 #include "shell/gpu/gpu_surface_vulkan_impeller.h"                // nogncheck
-
-static std::vector<std::shared_ptr<fml::Mapping>> ShaderLibraryMappings() {
-  return {
-      std::make_shared<fml::NonOwnedMapping>(impeller_entity_shaders_vk_data,
-                                             impeller_entity_shaders_vk_length),
-      std::make_shared<fml::NonOwnedMapping>(impeller_modern_shaders_vk_data,
-                                             impeller_modern_shaders_vk_length),
-      std::make_shared<fml::NonOwnedMapping>(
-          impeller_framebuffer_blend_shaders_vk_data,
-          impeller_framebuffer_blend_shaders_vk_length),
-      std::make_shared<fml::NonOwnedMapping>(
-          impeller_compute_shaders_vk_data, impeller_compute_shaders_vk_length),
-  };
-}
-
-struct ImpellerVulkanContextHolder {
-  ImpellerVulkanContextHolder() = default;
-  ImpellerVulkanContextHolder(ImpellerVulkanContextHolder&&) = default;
-  std::shared_ptr<impeller::ContextVK> context;
-  std::shared_ptr<impeller::SurfaceContextVK> surface_context;
-
-  bool Initialize(bool enable_validation);
-};
-
-bool ImpellerVulkanContextHolder::Initialize(bool enable_validation) {
-  impeller::ContextVK::Settings context_settings;
-  context_settings.proc_address_callback = &vkGetInstanceProcAddr;
-  context_settings.shader_libraries_data = ShaderLibraryMappings();
-  context_settings.cache_directory = fml::paths::GetCachesDirectory();
-  context_settings.enable_validation = enable_validation;
-  // Enable lazy shader mode for faster test execution as most tests
-  // will never render anything at all.
-  context_settings.flags.lazy_shader_mode = true;
-
-  context = impeller::ContextVK::Create(std::move(context_settings));
-  if (!context || !context->IsValid()) {
-    VALIDATION_LOG << "Could not create Vulkan context.";
-    return false;
-  }
-
-  impeller::vk::SurfaceKHR vk_surface;
-  impeller::vk::HeadlessSurfaceCreateInfoEXT surface_create_info;
-  auto res = context->GetInstance().createHeadlessSurfaceEXT(
-      &surface_create_info,  // surface create info
-      nullptr,               // allocator
-      &vk_surface            // surface
-  );
-  if (res != impeller::vk::Result::eSuccess) {
-    VALIDATION_LOG << "Could not create surface for tester "
-                   << impeller::vk::to_string(res);
-    return false;
-  }
-
-  impeller::vk::UniqueSurfaceKHR surface{vk_surface, context->GetInstance()};
-  surface_context = context->CreateSurfaceContext();
-  if (!surface_context->SetWindowSurface(std::move(surface),
-                                         impeller::ISize{1, 1})) {
-    VALIDATION_LOG << "Could not set up surface for context.";
-    return false;
-  }
-  return true;
-}
-
-#else
-struct ImpellerVulkanContextHolder {};
-#endif  // IMPELLER_SUPPORTS_RENDERING
+#endif  // TESTER_ENABLE_VULKAN
 
 #if defined(FML_OS_WIN)
 #include <combaseapi.h>
@@ -196,7 +124,7 @@ class TesterPlatformView : public PlatformView,
         impeller_context_holder_(std::move(impeller_context_holder)) {}
 
   ~TesterPlatformView() {
-#if ALLOW_IMPELLER
+#if TESTER_ENABLE_VULKAN
     if (impeller_context_holder_.context) {
       impeller_context_holder_.context->Shutdown();
     }
@@ -205,17 +133,17 @@ class TesterPlatformView : public PlatformView,
 
   // |PlatformView|
   std::shared_ptr<impeller::Context> GetImpellerContext() const override {
-#if ALLOW_IMPELLER
+#if TESTER_ENABLE_VULKAN
     return std::static_pointer_cast<impeller::Context>(
         impeller_context_holder_.context);
 #else
     return nullptr;
-#endif  // ALLOW_IMPELLER
+#endif  // TESTER_ENABLE_VULKAN
   }
 
   // |PlatformView|
   std::unique_ptr<Surface> CreateRenderingSurface() override {
-#if ALLOW_IMPELLER
+#if TESTER_ENABLE_VULKAN
     if (delegate_.OnPlatformViewGetSettings().enable_impeller) {
       FML_DCHECK(impeller_context_holder_.context);
       auto surface = std::make_unique<GPUSurfaceVulkanImpeller>(
@@ -223,7 +151,7 @@ class TesterPlatformView : public PlatformView,
       FML_DCHECK(surface->IsValid());
       return surface;
     }
-#endif  // ALLOW_IMPELLER
+#endif  // TESTER_ENABLE_VULKAN
     auto surface = std::make_unique<TesterGPUSurfaceSoftware>(
         this, true /* render to surface */);
     FML_DCHECK(surface->IsValid());
@@ -385,14 +313,14 @@ int RunTester(const flutter::Settings& settings,
 
   ImpellerVulkanContextHolder impeller_context_holder;
 
-#if ALLOW_IMPELLER
+#if TESTER_ENABLE_VULKAN
   if (settings.enable_impeller) {
     if (!impeller_context_holder.Initialize(
             settings.enable_vulkan_validation)) {
       return EXIT_FAILURE;
     }
   }
-#endif  // ALLOW_IMPELLER
+#endif  // TESTER_ENABLE_VULKAN
 
   Shell::CreateCallback<PlatformView> on_create_platform_view =
       fml::MakeCopyable([impeller_context_holder = std::move(
